requirements.txt
Flask
Flask-OAuthlib
gunicorn
spotipy
authlib
tensorflow
keras
python-dotenv